llamaindex-projects/VectorSync/requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.13
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
attrs==25.1.0
azure-ai-inference==1.0.0b9
azure-core==1.32.0
azure-identity==1.20.0
azure-storage-blob==12.24.1
beautifulsoup4==4.13.3
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.2
dataclasses-json==0.6.7
debugpy @ file:///D:/bld/debugpy_1737269745117/work
Deprecated==1.2.18
dirtyjson==1.0.8
distro==1.9.0
filetype==1.2.0
frozenlist==1.5.0
fsspec==2025.2.0
googleapis-common-protos==1.69.1
greenlet==3.1.1
grpcio==1.70.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
isodate==0.7.2
jiter==0.8.2
joblib==1.4.2
llama-cloud==0.1.13
llama-cloud-services==0.6.3
llama-index==0.10.68
llama-index-agent-openai==0.2.9
llama-index-cli==0.1.13
llama-index-core==0.10.68.post1
llama-index-embeddings-azure-inference==0.3.0
llama-index-embeddings-openai==0.1.11
llama-index-indices-managed-llama-cloud==0.2.7
llama-index-legacy==0.9.48.post4
llama-index-llms-azure-inference==0.3.0
llama-index-llms-azure-openai==0.3.0
llama-index-llms-openai==0.1.31
llama-index-multi-modal-llms-openai==0.1.9
llama-index-program-openai==0.1.7
llama-index-question-gen-openai==0.1.3
llama-index-readers-file==0.1.33
llama-index-readers-llama-parse==0.1.6
llama-index-readers-microsoft-sharepoint==0.5.0
llama-index-vector-stores-pinecone==0.1.2
llama-parse==0.4.9
lz4==4.4.3
marshmallow==3.26.1
msal==1.31.1
msal-extensions==1.2.0
multidict==6.1.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
numpy==1.26.4
openai==1.65.2
packaging==24.2
pandas==2.2.3
pillow==11.1.0
pinecone==6.0.1
pinecone-client==3.2.2
pinecone-plugin-interface==0.0.7
portalocker==2.10.1
propcache==0.3.0
protobuf==5.29.3
protoc-gen-openapiv2==0.0.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.10.1
pypdf==4.3.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
pywin32==308
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
setuptools==75.8.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.38
striprtf==0.0.26
tenacity==8.5.0
tiktoken==0.9.0
tqdm==4.67.1
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
wheel==0.45.1
wrapt==1.17.2
yarl==1.18.3
